chore(figma): rename endpoint/webhook files to kebab-case
