+++ packages/figma/webhooks/file-delete.ts
